Wrapped path in new class Path
The path stored in PathFollower is now wrapped in a Path instance which
merges and extends PathWithPosition and PathFollower::PathIndex, making
both obsolete.
Switching to next waypoint or sub path is completely managed by this new
class which makes the code more clear and prevents other modules to do
ugly things with the path position. Furthermore the encapsulation of the
waypoint index is necessary for the not yet implemented
next-waypoint-event of the supervisor chain.