Separate enum for status of computeMoveCommand()
As the ControlStatus contains a OBSTACLE status but computeMoveCommand()
does not look for obstacles, a new enum MoveCommandStatus was created,
which is a subset of the ControlStatus.
Not really a nice solution but I think it is cleaner this way, as
computeMoveCommand() is prevented from doing weird stuff with the return
status.