Use `*_NAMED` macros for output of submodules
All submodules (controller, obstacle avoider, ...) are using the "named"
versions of the rosconsole macros (e.g. `ROS_INFO_NAMED` instead of
`ROS_INFO`). This has the advantage, that the output level of those
submodules can be controlled independently. For example when debugging
one module, the debug-output of all the other nodules can be
deactivated.