Better obstacle clustering in `PathLookout`
The old clustering algorithm assumed that the source of the points one
laser scanner and that the points are ordered by angle. This assumption
is not true for arbitrary sources of obstacle clouds (e.g. 3d sensors or
if the cloud is a combined result of multiple sensors).
Therefore the clustering was replaced by a new greedy algorithm that
should yield good enought results for the given purpose while still
being significantly faster than a more complex algorithm like k-means.