Started to implement new PID controller
The new PID controller implemented in `PidController` implements a full
PID controller (old `PidCtrl` is actually only PI) that works with
matrices and can thus be used to controll multidimensional values
(discarding the need for `MultiplePidWrapper`).
The new controller class is not yet ready to use, this is merely a
"backup commit".