Fixes #15: Wait for transform + handle fail
Wait some small time (currently 10 ms) for a transformation and abort, if non was received. This aviods calling `transformPointCloud` when there is no transformation available (which will result in a error message that spams in the terminal).
Further fixed the way how failing transformations are handled. Different to the corresponding tf methods, `pcl_ros::transformPointCloud` does not throw an exception but return `false` if something fails.